Работает всё кроме отчета

+++ Project1/Project1/MainForm.cs
@@ -114,33 +114,57 @@ namespace Project1
             {
                 case 0:
                     int id = 0;
-                    for (int i = 0; i < dataGridView1.SelectedRows.Count; i++)
+                    try
                     {
-                        int index = dataGridView1.SelectedRows[0].Index;
-                        Int32.TryParse(dataGridView1[0, index].Value.ToString(), out id);
-                        PeripheralDevice player = db.PeripheralDevices.Find(id);
-                        db.PeripheralDevices.Remove(player);
+                        for (int i = 0; i < dataGridView1.RowCount; i++)
+                        {
+                            if (dataGridView1.Rows[i].Selected == true)
+                            {
+                                Int32.TryParse(dataGridView1[0, i].Value.ToString(), out id);
+                                PeripheralDevice player = db.PeripheralDevices.Find(id);
+                                db.PeripheralDevices.Remove(player);
+                            }
+                        }
+                        db.SaveChanges();
+                        UpdatePD();
                     }
-                    db.SaveChanges();
-                    UpdatePD();
+                    catch { MessageBox.Show("Надо выбрать строку полностью"); }
                     break;
                 case 1:
-                    int index1 = dataGridView2.SelectedRows[0].Index;
                     int id1 = 0;
-                    Int32.TryParse(dataGridView2[0, index1].Value.ToString(), out id1);
-                    Computer player1 = db.Computers.Find(id1);
-                    db.Computers.Remove(player1);
-                    db.SaveChanges();
-                    UpdateComp();
+                    try
+                    {
+                        for (int i = 0; i < dataGridView2.RowCount; i++)
+                        {
+                            if (dataGridView2.Rows[i].Selected == true)
+                            {
+                                Int32.TryParse(dataGridView2[0, i].Value.ToString(), out id1);
+                                Computer player1 = db.Computers.Find(id1);
+                                db.Computers.Remove(player1);
+                            }
+                        }
+                        db.SaveChanges();
+                        UpdateComp();
+                    }
+                    catch { MessageBox.Show("Надо выбрать строку полностью"); }
                     break;
                 case 2:
-                    int index2 = dataGridView3.SelectedRows[0].Index;
                     int id2 = 0;
-                    Int32.TryParse(dataGridView3[0, index2].Value.ToString(), out id2);
-                    Employee player2 = db.Employees.Find(id2);
-                    db.Employees.Remove(player2);
-                    db.SaveChanges();
-                    UpdateEmployee();
+                    try
+                    {
+                        for (int i = 0; i < dataGridView3.RowCount; i++)
+                        {
+                            if (dataGridView2.Rows[i].Selected == true)
+                            {
+                                Int32.TryParse(dataGridView3[0, i].Value.ToString(), out id2);
+                                Employee player2 = db.Employees.Find(id2);
+                                db.Employees.Remove(player2);
+                            }
+                        }
+                        db.SaveChanges();
+                        UpdateEmployee();
+                    }
+                    catch { MessageBox.Show("Надо выбрать строку полностью"); }
                     break;
             }
         }
@@ -172,111 +196,122 @@ namespace Project1
             switch (Tabform.SelectedIndex)
             {
                 case 0:
-                    int index = dataGridView1.SelectedRows[0].Index;
-                    int id = 0;
-                    Int32.TryParse(dataGridView1[0, index].Value.ToString(), out id);
+                    try
+                    {
+                        int index = dataGridView1.SelectedRows[0].Index;
+                        int id = 0;
+                        Int32.TryParse(dataGridView1[0, index].Value.ToString(), out id);
 
-                    PeripheralDevice player = db.PeripheralDevices.Find(id);
+                        PeripheralDevice player = db.PeripheralDevices.Find(id);
 
-                    Redit plForm = new Redit();
+                        Redit plForm = new Redit();
 
-                    plForm.comboBox4.SelectedItem = player.status;
-                    plForm.comboBox1.SelectedValue = player.ManufacturerId;
-                    plForm.textBox2.Text = player.name;
-                    plForm.comboBox2.SelectedValue = player.TypeId;
-                    plForm.comboBox3.SelectedValue = player.DepartamentId;
-                    plForm.comboBox5.SelectedValue = player.ComputerId;
+                        plForm.comboBox4.SelectedItem = player.status;
+                        plForm.comboBox1.SelectedValue = player.ManufacturerId;
+                        plForm.textBox2.Text = player.name;
+                        plForm.comboBox2.SelectedValue = player.TypeId;
+                        plForm.comboBox3.SelectedValue = player.DepartamentId;
+                        plForm.comboBox5.SelectedValue = player.ComputerId;
 
-                    DialogResult result = plForm.ShowDialog(this);
+                        DialogResult result = plForm.ShowDialog(this);
 
-                    if (result == DialogResult.Cancel)
-                        return;
+                        if (result == DialogResult.Cancel)
+                            return;
 
-                    player.status = plForm.comboBox4.SelectedItem.ToString();
-                    player.name = plForm.textBox2.Text;
+                        player.status = plForm.comboBox4.SelectedItem.ToString();
+                        player.name = plForm.textBox2.Text;
 
-                    var val = int.Parse(plForm.comboBox1.SelectedValue.ToString());
-                    player.manufacturer = db.Manufacturers.Where(s => s.ManufacturerId == val).FirstOrDefault<Manufacturer>();
+                        var val = int.Parse(plForm.comboBox1.SelectedValue.ToString());
+                        player.manufacturer = db.Manufacturers.Where(s => s.ManufacturerId == val).FirstOrDefault<Manufacturer>();
 
-                    var val1 = int.Parse(plForm.comboBox2.SelectedValue.ToString());
-                    player.type = db.Types.Where(s => s.TypeId == val1).FirstOrDefault<Type>();
+                        var val1 = int.Parse(plForm.comboBox2.SelectedValue.ToString());
+                        player.type = db.Types.Where(s => s.TypeId == val1).FirstOrDefault<Type>();
 
-                    var val12 = int.Parse(plForm.comboBox3.SelectedValue.ToString());
-                    player.departament = db.Departaments.Where(s => s.DepartamentId == val12).FirstOrDefault<Departament>();
+                        var val12 = int.Parse(plForm.comboBox3.SelectedValue.ToString());
+                        player.departament = db.Departaments.Where(s => s.DepartamentId == val12).FirstOrDefault<Departament>();
 
-                    var val3 = int.Parse(plForm.comboBox5.SelectedValue.ToString());
-                    player.computer = db.Computers.Where(s => s.ComputerId == val3).FirstOrDefault<Computer>();
+                        var val3 = int.Parse(plForm.comboBox5.SelectedValue.ToString());
+                        player.computer = db.Computers.Where(s => s.ComputerId == val3).FirstOrDefault<Computer>();
 
-                    db.SaveChanges();
-                    UpdatePD();
-                    MessageBox.Show("Объект обновлен");
+                        db.SaveChanges();
+                        UpdatePD();
+                        MessageBox.Show("Объект обновлен");
+                    }
+                    catch { MessageBox.Show("Надо выбрать строку полностью"); }
                     break;
-
                 case 1:
-                    int index1 = dataGridView2.SelectedRows[0].Index;
-                    int id1 = 0;
-                    Int32.TryParse(dataGridView2[0, index1].Value.ToString(), out id1);
+                    try
+                    {
+                        int index1 = dataGridView2.SelectedRows[0].Index;
+                        int id1 = 0;
+                        Int32.TryParse(dataGridView2[0, index1].Value.ToString(), out id1);
 
-                    Computer player1 = db.Computers.Find(id1);
+                        Computer player1 = db.Computers.Find(id1);
 
-                    CEdit plForm1 = new CEdit();
+                        CEdit plForm1 = new CEdit();
 
-                    plForm1.comboBox4.SelectedItem = player1.status;
-                    plForm1.comboBox1.SelectedValue = player1.MotherboardId;
-                    plForm1.comboBox2.SelectedValue = player1.HddId;
-                    plForm1.comboBox3.SelectedValue = player1.CpuId;
+                        plForm1.comboBox4.SelectedItem = player1.status;
+                        plForm1.comboBox1.SelectedValue = player1.MotherboardId;
+                        plForm1.comboBox2.SelectedValue = player1.HddId;
+                        plForm1.comboBox3.SelectedValue = player1.CpuId;
 
-                    DialogResult result1 = plForm1.ShowDialog(this);
+                        DialogResult result1 = plForm1.ShowDialog(this);
 
-                    if (result1 == DialogResult.Cancel)
-                        return;
+                        if (result1 == DialogResult.Cancel)
+                            return;
 
-                    player1.status = plForm1.comboBox4.SelectedItem.ToString();
+                        player1.status = plForm1.comboBox4.SelectedItem.ToString();
 
-                    var val5 = int.Parse(plForm1.comboBox1.SelectedValue.ToString());
-                    player1.motherboard = db.Motherboards.Where(s => s.MotherboardId == val5).FirstOrDefault<Motherboard>();
+                        var val5 = int.Parse(plForm1.comboBox1.SelectedValue.ToString());
+                        player1.motherboard = db.Motherboards.Where(s => s.MotherboardId == val5).FirstOrDefault<Motherboard>();
 
-                    var val6 = int.Parse(plForm1.comboBox2.SelectedValue.ToString());
-                    player1.hdd = db.HDDs.Where(s => s.HddId == val6).FirstOrDefault<HDD>();
+                        var val6 = int.Parse(plForm1.comboBox2.SelectedValue.ToString());
+                        player1.hdd = db.HDDs.Where(s => s.HddId == val6).FirstOrDefault<HDD>();
 
-                    var val7 = int.Parse(plForm1.comboBox3.SelectedValue.ToString());
-                    player1.cpu = db.CPUs.Where(s => s.CpuId == val7).FirstOrDefault<CPU>();
+                        var val7 = int.Parse(plForm1.comboBox3.SelectedValue.ToString());
+                        player1.cpu = db.CPUs.Where(s => s.CpuId == val7).FirstOrDefault<CPU>();
 
-                    db.SaveChanges();
-                    UpdateComp();
-                    MessageBox.Show("Объект обновлен");
+                        db.SaveChanges();
+                        UpdateComp();
+                        MessageBox.Show("Объект обновлен");
+                    }
+                    catch { MessageBox.Show("Надо выбрать строку полностью"); }
                     break;
 
                 case 2:
-                    int index2 = dataGridView3.SelectedRows[0].Index;
-                    int id2 = 0;
-                    Int32.TryParse(dataGridView3[0, index2].Value.ToString(), out id2);
+                    try
+                    {
+                        int index2 = dataGridView3.SelectedRows[0].Index;
+                        int id2 = 0;
+                        Int32.TryParse(dataGridView3[0, index2].Value.ToString(), out id2);
 
-                    Employee player2 = db.Employees.Find(id2);
+                        Employee player2 = db.Employees.Find(id2);
 
-                    Eedit plForm2 = new Eedit();
+                        Eedit plForm2 = new Eedit();
 
-                    plForm2.textBox1.Text = player2.fullName;
-                    plForm2.textBox2.Text = player2.position;
-                    plForm2.comboBox1.SelectedValue = player2.DepartamentId;
-                    plForm2.comboBox2.SelectedValue = player2.ComputerId;
+                        plForm2.textBox1.Text = player2.fullName;
+                        plForm2.textBox2.Text = player2.position;
+                        plForm2.comboBox1.SelectedValue = player2.DepartamentId;
+                        plForm2.comboBox2.SelectedValue = player2.ComputerId;
 
-                    DialogResult result2 = plForm2.ShowDialog(this);
+                        DialogResult result2 = plForm2.ShowDialog(this);
 
-                    if (result2 == DialogResult.Cancel)
-                        return;
+                        if (result2 == DialogResult.Cancel)
+                            return;
 
-                    player2.fullName = plForm2.textBox1.Text;
-                    player2.position = plForm2.textBox2.Text;
-                    var val8 = int.Parse(plForm2.comboBox1.SelectedValue.ToString());
-                    player2.departament = db.Departaments.Where(s => s.DepartamentId == val8).FirstOrDefault<Departament>();
+                        player2.fullName = plForm2.textBox1.Text;
+                        player2.position = plForm2.textBox2.Text;
+                        var val8 = int.Parse(plForm2.comboBox1.SelectedValue.ToString());
+                        player2.departament = db.Departaments.Where(s => s.DepartamentId == val8).FirstOrDefault<Departament>();
 
-                    var val9 = int.Parse(plForm2.comboBox2.SelectedValue.ToString());
-                    player2.computer = db.Computers.Where(s => s.ComputerId == val9).FirstOrDefault<Computer>();
+                        var val9 = int.Parse(plForm2.comboBox2.SelectedValue.ToString());
+                        player2.computer = db.Computers.Where(s => s.ComputerId == val9).FirstOrDefault<Computer>();
 
-                    db.SaveChanges();
-                    UpdateEmployee();
-                    MessageBox.Show("Объект обновлен");
+                        db.SaveChanges();
+                        UpdateEmployee();
+                        MessageBox.Show("Объект обновлен");
+                    }
+                    catch { MessageBox.Show("Надо выбрать строку полностью"); }
                     break;
             }
         }
